Backend/requirements.txt
# Backend requirements
fastapi
uvicorn
gunicorn # For production deployment
PyMuPDF
python-dotenv
google-cloud-translate
reportlab
firebase-admin
google-generativeai
networkx
firebase-admin
python-dotenv
